AutonomousAINewsChannel/python-news-service/requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
langgraph
langchain